+++ src/Common/Common.Api/IPowerBIApiClient.cs
@@ -7,7 +7,6 @@ using System;
 using Microsoft.PowerBI.Common.Api.Datasets;
 using Microsoft.PowerBI.Common.Api.Reports;
 using Microsoft.PowerBI.Common.Api.Workspaces;
-using Microsoft.PowerBI.Common.Api.Imports;
 
 namespace Microsoft.PowerBI.Common.Api
 {
@@ -18,7 +17,6 @@ namespace Microsoft.PowerBI.Common.Api
         IWorkspacesClient Workspaces { get; set; }
 
         IDatasetsClient Datasets { get; set; }
-
-        IImportsClient Imports { get; set; }
+      
     }
 }
