--- src/Common/Common.Api/Imports/IImportsClient.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.PowerBI.Api.V2.Models;
-
-namespace Microsoft.PowerBI.Common.Api.Imports
-{
-    public interface IImportsClient
-    {
-        Guid PostImport(string datasetDisplayName, string filePath);
-    }
-}
