+++ src/Common/Common.Api/PowerBIApiClient.cs
@@ -8,7 +8,6 @@ using System.Net.Http;
 using Microsoft.PowerBI.Api.V2;
 using Microsoft.PowerBI.Common.Abstractions.Interfaces;
 using Microsoft.PowerBI.Common.Api.Datasets;
-using Microsoft.PowerBI.Common.Api.Imports;
 using Microsoft.PowerBI.Common.Api.Reports;
 using Microsoft.PowerBI.Common.Api.Workspaces;
 using Microsoft.Rest;
@@ -26,9 +25,7 @@ namespace Microsoft.PowerBI.Common.Api
         public IWorkspacesClient Workspaces { get; set; }
 
         public IDatasetsClient Datasets { get; set; }
-
-        public IImportsClient Imports { get; set; }
-
+        
         public PowerBIApiClient(IAuthenticationFactory authenticator, IPowerBIProfile profile, IPowerBILogger logger, IPowerBISettings settings)
         {
             this.Client = CreateClient(authenticator, profile, logger, settings);
@@ -46,7 +43,6 @@ namespace Microsoft.PowerBI.Common.Api
             this.Reports = new ReportsClient(this.Client);
             this.Workspaces = new WorkspacesClient(this.Client);
             this.Datasets = new DatasetsClient(this.Client);
-            this.Imports = new ImportsClient(this.Client);
         }
 
         private static IPowerBIClient CreateClient(IAuthenticationFactory authenticator, IPowerBIProfile profile, IPowerBILogger logger, IPowerBISettings settings)
@@ -117,11 +113,6 @@ namespace Microsoft.PowerBI.Common.Api
                 if (this.Datasets != null)
                 {
                     this.Datasets = null;
-                }
-
-                if (this.Imports != null)
-                {
-                    this.Imports = null;
                 }
             }
         }
